+++ lib/playground/application.ex
@@ -22,6 +22,9 @@ defmodule Playground.Application do
       Playground.Engine
     ]
 
+    # Capture error messages coming from process crashes in Sentry
+    :logger.add_handler(:sentry_handler, Sentry.LoggerHandler, %{})
+
     # See https://hexdocs.pm/elixir/Supervisor.html
     # for other strategies and supported options
     opts = [strategy: :one_for_one, name: Playground.Supervisor]