+++ lib/playground_web.ex
@@ -54,6 +54,8 @@ defmodule PlaygroundWeb do
       use Phoenix.LiveView,
         layout: {PlaygroundWeb.Layouts, :app}
 
+      on_mount Sentry.LiveViewHook
+
       unquote(html_helpers())
     end
   end