+++ lib/playground_web/endpoint.ex
@@ -1,4 +1,5 @@
 defmodule PlaygroundWeb.Endpoint do
+  use Sentry.PlugCapture
   use Phoenix.Endpoint, otp_app: :playground
 
   # The session will be stored in the cookie and signed,
@@ -47,6 +48,8 @@ defmodule PlaygroundWeb.Endpoint do
     parsers: [:urlencoded, :multipart, :json],
     pass: ["*/*"],
     json_decoder: Phoenix.json_library()
+
+  plug Sentry.PlugContext
 
   plug Plug.MethodOverride
   plug Plug.Head