+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Super tic tac toe: timer of current move + cumulative of each player
+- Add sentry for error capturing
 
 ### Changed